requirements.txt
# Dependencias principales
streamlit==1.45.1
mysql-connector-python==9.3.0
bcrypt==4.1.2
PyJWT==2.8.0

# Dependencias de datos y análisis
pandas==2.2.3
numpy==2.2.6
python-dateutil==2.9.0.post0

# Dependencias de interfaz
streamlit-option-menu==0.3.12
pillow==11.2.1

# Dependencias de desarrollo y testing
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0

# Dependencias de logging y configuración
python-dotenv==1.0.0
pydantic==2.5.3

# Dependencias de seguridad
cryptography==41.0.8

# Dependencias de validación
email-validator==2.1.0

# Dependencias de serialización
marshmallow==3.20.1

# Dependencias de caché
redis==5.0.1

# Dependencias de monitoreo
structlog==24.1.0

# Dependencias de documentación
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
